Exposed get_env_var
