+++ src/find_library.cpp
@@ -25,6 +25,8 @@
 #include "rcutils/filesystem.h"
 #include "rcutils/get_env.h"
 
+#include "rcpputils/get_env.hpp"
+
 namespace rcpputils
 {
 
@@ -47,16 +49,6 @@ static constexpr char kPathSeparator = ':';
 static constexpr char kSolibPrefix[] = "lib";
 static constexpr char kSolibExtension[] = ".so";
 #endif
-
-std::string get_env_var(const char * env_var)
-{
-  const char * value{};
-  const char * err = rcutils_get_env(env_var, &value);
-  if (err) {
-    throw std::runtime_error(err);
-  }
-  return value ? value : "";
-}
 
 std::list<std::string> split(const std::string & value, const char delimiter)
 {
